backend/requirements.txt
Authlib==1.6.0
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
Flask==3.1.1
flask-cors==6.0.1
Flask-Session==0.8.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
psycopg2==2.9.10
pycparser==2.22
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
redis==6.2.0
requests==2.32.4
simple-websocket==1.1.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
